+++ build.js
@@ -55,6 +55,7 @@ const appJs = [
   'src/js/ui.js',
   'src/js/filters.js',
   'src/js/modal.js',
+  'src/js/modal-field-controls.js',
   'src/js/modal-chart-fields.js',
   'src/js/modal-fields.js',
   'src/js/date-prompt.js',